// DriverScore_frontend-main/src/client/models/ImuSampleSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ImuSampleSchema = {
    timestamp: string;
    acceleration_x_ms2: number;
    acceleration_y_ms2: number;
    acceleration_z_ms2: number;
    angular_velocity_x_rads: number;
    angular_velocity_y_rads: number;
    angular_velocity_z_rads: number;
    pitch_rad: number;
    yaw_rad: number;
    roll_rad: number;
};